% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGI0.R
\name{dGI0}
\alias{dGI0}
\title{Probability density function of the GI0 distribution}
\usage{
dGI0(x, p_alpha, p_gamma, p_Looks)
}
\arguments{
\item{p_alpha}{Negative value that controls roughness}

\item{p_gamma}{Positive value that controls scale}

\item{p_Looks}{Positive value that controls number of looks}
}
\value{
Density value (numeric)
}
\description{
This function returns the density of point \eqn{x} on a GI0 distribution with parameters \eqn{\alpha}, \eqn{\gamma} and \eqn{Looks}.
}
\examples{
dGI0(5, -2, 3, 4) # -> 0.0584
}
\keyword{GI0}
\keyword{density}
\keyword{distribution}
