% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pGI0.R
\name{pGI0}
\alias{pGI0}
\title{Cumulative distribution function of the GI0 distribution}
\usage{
pGI0(x, p_alpha, p_gamma, p_Looks, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p_alpha}{Required: negative value that controls roughness}

\item{p_gamma}{Required: positive value that controls scale}

\item{p_Looks}{Required: positive value greater or equal to 1 that controls number of looks}

\item{lower.tail}{Logical. If \code{TRUE}, probabilities are \eqn{P(X \leq x)}; otherwise, \eqn{P(X > x)}}

\item{log.p}{Logical. If \code{TRUE}, probabilities are stated as \eqn{\log(p)}}
}
\value{
Distribution function value (numeric)
}
\description{
This function returns the cumulative probability of point \eqn{x} on a GI0 distribution with parameters \eqn{\alpha}, \eqn{\gamma}
and \eqn{Looks}.
}
\examples{
pGI0(3.2, -2, 3, 7) # -> 0.7579
}
\keyword{GI0}
\keyword{cumulative}
\keyword{distribution}
