% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qGI0.R
\name{qGI0}
\alias{qGI0}
\title{Quantile function of the GI0 distribution}
\usage{
qGI0(p, p_alpha, p_gamma, p_Looks, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p_alpha}{Negative value that controls roughness}

\item{p_gamma}{Positive value that controls scale}

\item{p_Looks}{Positive value that controls number of looks}

\item{lower.tail}{Logical. If \code{TRUE}, values are \eqn{X \leq Q(p)}; otherwise, \eqn{X > Q(p)}}

\item{log.p}{Logical. If \code{TRUE}, probabilities are stated as \eqn{\log(p)}}
}
\value{
Quantile function value (numeric)
}
\description{
This function returns the quantile of point \eqn{x} on a GI0 distribution with parameters \eqn{\alpha}, \eqn{\gamma} and \eqn{Looks}.
}
\examples{
qGI0(0.34, -2, 5, 4) # -> 0.7555
}
\keyword{GI0}
\keyword{distribution}
\keyword{quantile}
