% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGI0.R
\name{rGI0}
\alias{rGI0}
\title{Random variate generator from the GI0 distribution}
\usage{
rGI0(n, p_alpha, p_gamma, p_Looks, from.F = FALSE)
}
\arguments{
\item{n}{Sample size to generate.}

\item{p_alpha}{Negative value that controls roughness}

\item{p_gamma}{Positive value that controls scale}

\item{p_Looks}{Positive value that controls number of looks}

\item{from.F}{Logical. If \code{TRUE}, generate from F-distribution; otherwise, generate as ratio of two Gamma variates}
}
\value{
Vector of random variates with length \eqn{n}
}
\description{
This function returns \eqn{n} random variates from a GI0 distribution with parameters \eqn{\alpha}, \eqn{\gamma} and \eqn{Looks}.
}
\examples{
set.seed(1967, kind="Mersenne-Twister")
rGI0(5, -2, 4, 5) # -> returns (0.535, 0.842 2.635 3.863 1.599)
}
\keyword{GI0}
\keyword{distribution}
\keyword{random}
